+++ src/main/java/com/supermartijn642/fusion/predicate/MatchBlockConnectionPredicate.java
@@ -25,7 +25,7 @@ public class MatchBlockConnectionPredicate implements ConnectionPredicate {
                 throw new JsonParseException("Match block predicate must have string property 'block'!");
             if(!IdentifierUtil.isValidIdentifier(json.get("block").getAsString()))
                 throw new JsonParseException("Property 'block' must be a valid identifier!");
-            ResourceLocation identifier = new ResourceLocation(json.get("block").getAsString());
+            ResourceLocation identifier = ResourceLocation.parse(json.get("block").getAsString());
             if(!BuiltInRegistries.BLOCK.containsKey(identifier))
                 throw new JsonParseException("Unknown block '" + identifier + "'!");
             Block block = BuiltInRegistries.BLOCK.get(identifier);