+++ changelog.md
@@ -1,2 +1,2 @@
 ### Fusion 1.1.1
-- Initial release of Fusion for Minecraft 1.20.5 & 1.20.6
+- Initial release of Fusion for Minecraft 1.21 & 1.21.1